Moved updater business logic in updater package
